+++ nvim/lua/plugins/ui.lua
@@ -137,7 +137,8 @@ return {
       {
         "<leader>e",
         function()
-          require("neo-tree.command").execute({ toggle = true, dir = vim.loop.cwd() })
+          local uv = vim.uv or vim.loop
+          require("neo-tree.command").execute({ toggle = true, dir = uv.cwd() })
         end,
         desc = "Explorer NeoTree (cwd)",
       },
@@ -148,7 +149,8 @@ return {
     end,
     init = function()
       if vim.fn.argc(-1) == 1 then
-        local stat = vim.loop.fs_stat(vim.fn.argv(0))
+        local uv = vim.uv or vim.loop
+        local stat = uv.fs_stat(vim.fn.argv(0))
         if stat and stat.type == "directory" then
           require("neo-tree")
         end